+++ app/Http/Controllers/PostController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Exceptions\ModelNotFoundException;
 use App\Post;
 use App\Report;
 use App\Category;
+use App\Admin;
 
 class PostController extends Controller
 {
@@ -25,6 +27,10 @@ class PostController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
+
+        if($post == NULL)
+            throw new ModelNotFoundException;
+
         $comments = Post::with('comments.user')->find($id);
         $comments = $comments->comments;
         return view('PostView.show',compact('post','comments'));
@@ -55,6 +61,11 @@ class PostController extends Controller
     public function delete(Request $request)
     {
         $post = Post::find($request->id);
+
+        $admin = Admin::where('user_id','=',\Auth::id())->get();
+
+        if(\Auth::id()!= $post->user_id && $admin->isEmpty())
+            abort(500);
 
         $post->delete();
